Move Hydra CLI into packaged source tree
