+++ codex_ml/cli/main.py
@@ -1,50 +1,24 @@
+"""Shim to src/codex_ml/cli/main.py.
+
+This keeps local imports in sync with the packaged module when the repository
+root shadows the installed package on ``sys.path``.
+"""
+
 from __future__ import annotations
 
-import sys
+import importlib.util
+import pathlib
 
-import hydra
-from omegaconf import DictConfig, OmegaConf
-
-try:  # best-effort imports; functions are optional
-    from codex_ml.trainer import run_training  # type: ignore
-except Exception:  # pragma: no cover
-
-    def run_training(cfg):  # type: ignore
-        return None
-
-
-try:
-    from codex_ml.eval.eval_runner import evaluate_datasets  # type: ignore
-except Exception:  # pragma: no cover
-
-    def evaluate_datasets(*args, **kwargs):  # type: ignore
-        return None
-
-
-@hydra.main(version_base="1.3", config_path="../../configs", config_name="config")
-def main(cfg: DictConfig) -> None:  # pragma: no cover - simple dispatcher
-    """Hydra CLI for Codex ML."""
-    print(OmegaConf.to_yaml(cfg))
-    for step in cfg.pipeline.steps:
-        if step == "train":
-            run_training(cfg.train)
-        elif step == "evaluate":
-            eval_cfg = OmegaConf.select(cfg, "eval")
-            if eval_cfg is None:
-                print("Eval config not found; skipping evaluate step", file=sys.stderr)
-                continue
-            datasets = eval_cfg.get("datasets", [])
-            metrics = eval_cfg.get("metrics", [])
-            output_dir = cfg.get("output_dir", "runs/eval")
-            evaluate_datasets(datasets, metrics, output_dir)
-    sys.exit(0)
-
-
-def cli(argv: list[str] | None = None) -> None:
-    if argv is not None:
-        sys.argv = [sys.argv[0]] + list(argv)
-    main()
-
-
-if __name__ == "__main__":  # pragma: no cover
-    cli()
+_src = (
+    pathlib.Path(__file__).resolve().parents[2]
+    / "src"
+    / "codex_ml"
+    / "cli"
+    / "main.py"
+)
+_spec = importlib.util.spec_from_file_location("codex_ml._src_cli_main", _src)
+_module = importlib.util.module_from_spec(_spec)
+assert _spec.loader is not None
+_spec.loader.exec_module(_module)
+__all__ = getattr(_module, "__all__", [])
+globals().update({k: v for k, v in _module.__dict__.items() if not k.startswith("_")})
